Import 2.3.42

+++ include/asm-m68k/checksum.h
@@ -123,7 +123,7 @@ ip_compute_csum(unsigned char * buff, int len)
 #define _HAVE_ARCH_IPV6_CSUM
 static __inline__ unsigned short int
 csum_ipv6_magic(struct in6_addr *saddr, struct in6_addr *daddr,
-		__u16 len, unsigned short proto, unsigned int sum) 
+		__u32 len, unsigned short proto, unsigned int sum) 
 {
 	register unsigned long tmp;
 	__asm__("addl %2@,%0\n\t"
@@ -145,7 +145,7 @@ csum_ipv6_magic(struct in6_addr *saddr, struct in6_addr *daddr,
 		"clrl %1\n\t"
 		"addxl %1,%0"
 		: "=&d" (sum), "=&d" (tmp)
-		: "a" (saddr), "a" (daddr), "d" ((__u32) len + proto),
+		: "a" (saddr), "a" (daddr), "d" (len + proto),
 		  "0" (sum));
 
 	return csum_fold(sum);
